slog: remove Attr.AppendValue
Attr.AppendValue was originally intended to be a convenient and more
efficient way of formatting Attr values. But there is no one clearly
right way to do format them, and the efficiency argument doesn't
really hold water: AppendValue skips the redundant but cheap type
check that occurs in

	switch a.Kind() {
	case IntKind:
		... a.Int() ..

but then formats Times and Durations as strings instead of integers.

We keep it internally to help out in a couple of places, but overall
it's better if Handler implementations do their own value formatting.
For casual use, fmt.Sprint(a.Value()) is fine.

Change-Id: Ie265f1ee5daf5f8002bfc2b70f4bf461cb07a10b