slog: both AtomicLevel and Level implement Leveler
Introduce the Leveler interface (only method: Level() Level) so that both
AtomicLevel and Level can be used interchangeably.

HandlerOptions.Level is now a Leveler.

It is tempting to get rid of NewAtomicLevel as unnecessary.
But if it's not there then someone might write

    var a AtomicLevel
    // I want the zero value so I won't set it.
    ... &a ...

The problem is that the zero value of AtomicLevel is 0, which does not
mean "enable all levels." With the current mapping of Levels to
numbers, it would actually mean "disable all positive levels." If we
made zero a special case, we'd have:

   1   enable levels <= 1
   0   enable all levels
   -1  enable levels <= -1

which would be confusing. Or we could insist that levels be
non-negative, but this would be the only place in the API that needs
that restriction.

If everyone sticks to *AtomicLevel, then the zero value is nil, which
we can interpret specially without confusion.

Change-Id: I0538201cffacd101901179a01f262ae6a5d96797