slog: rethink levels
We want the default level to be Info, so make it zero.

Then rejigger the level numbers to match OpenTelemetry:
- Reverse the direction, so larger numbers are more severe.
- Reduce the gaps between named levels from 10 to 4.

There is still a nice mapping to verbosities: negation.

By making the zero level valid, we lose a level value that means "no
level." We could reinstate it by picking an extreme value,
like math.MinInt. Or we could just omit it; perhaps it's
not needed.

Change-Id: I0f06418b7de79fde0c172e5c9d88c707fac74601