slog: split Value off from Attr
Move the value part of an Attr into its own type.

Besides making sense on general grounds--the key and
value are really independent of each other--it also
paves the way for

    type Marshaler interface {
        MarshalLog() Value
    }

which will provide a way for types to replace their
values for logging.

We considered having MarshalLog return Attr and ignore the key,
but this seems better.

The main negative consequence of this split is the duplication of
constructor functions. We use the longer XXXValue for Values, since
most people will construct Attrs, not Values. Fast switching on
generic type parameters (https://go.dev/issue/45380) would ameliorate
this, because we would need only one constructor function per type.

Change-Id: I974c5dba1bb0fbcd35aaf8e19f3a81daac6e8094