slog: add Marshaler interface
Add the Marshaler interface, which lets types replace their values
with a different value for logging.

Value has a new Marshaler Kind, so that Handlers that write a
switch over kinds like

    switch v.Kind() {
    case slog.StringKind: ...
    ...
    default:
        panic("bad kind")
    }

will find out that they are missing a case.

Add the Value.Resolve method to call MarshalLog repeatedly (but not
indefinitely).

Change-Id: I0e1bcb2c4d8508bafc552613daeb96e3373ece83