slog: faster Record.Clone
Record.Clone no longer copies the Record's []Attr.

Previously, Record.Clone made a copy of the Record's Attr slice.
Since a Record's list of Attrs can only be appended to, we don't need
to copy the slice, only clip it (setting capacity to length). We
are effectively doing copy-on-write.

After this change, is Clone still the right name?

Change-Id: I10d86b2b5b17e503224fc9f9bb5c07e70e8fca36