slog: add Group Values
Introduce a kind of Value that is a tagged group of other Values,
represented by a []Attr.

Now a type can provide a MarshalLog method that returns multiple Attrs,
so it can represent a loggable version of itself to slog.Handlers in
a general way.

This CL just adds the implementation to Value.
Subsequent CLs will provide Handler implementations and tests.

Change-Id: Ib0204daf146f545bd20abe4e58664899c8ef45db