slog: add WithScope
Add Logger.WithScope and Handler.WithScope, which provide a way
to scope attribute key names to a specific Logger. For example,
if two packages p1 and p2 both use the key "k", but mean different things by
it (perhaps use different types), that presents a challenge
for backends that ingest logs. But if p1 is passed the Logger resulting
from

    logger.WithScope("p1")

and p2 is passed

    looger.WithScope("p2")

then the Handler will qualify "k" by "p1" or "p2", respectively,
to produce keys that do not conflict.

Logger.WithScope just calls Handler.WithScope, so the implementation
is up to the Handler. Some may prepend scopes to key names; others
may use a scope name as a key to some nested map of key-value pairs.

This CL only defines the WithScope methods. Implementations and
tests will be in forthcoming CLs.

Change-Id: I45a5d204284e3a46c347b4069bb49ce1f68e63ea