slog: handle removed built-in keys
Fix commonHandler so it works correctly when all built-in keys are
removed and there are preformatted attributes.
In that case, we need to add a separator between the
preformatted attributes and the rest.

Also, generalize the test for built-in handlers.  (Some features won't
prove useful until scope is introduced in a later CL).

Change-Id: If1fdeae34cc59da26f3d7a9bb1bb38c3ad14d366