slog: fix SetDefault deadlock
If the program does

    SetDefault(Default())

with the original default logger, we end up with deadlock as
the call to log.Output invokes handlerWriter.Write, which
then calls log.Output again while the first invocation holds
the log.Logger mutex.

Fix by not installing a handleWriter if SetDefault's argument has
a defaultHandler.

Change-Id: I0db8a10f313d0e9913a65a75e163d7024ec96589