slog: FromContext delivers context to Handler
This CL provides a way for a Handler to obtain a context, by using the
FromContext method to add it to the Logger it returns.

It is often asked that Handler.Handle be given access to the context
of the log line. The main reason is to allow logs to contain a trace
ID, so they can be correlated with tracing. Observability systems like
Open Telemetry store trace information in the context.

The usual suggestion is to add a context argument to log methods:

    slog.Info(ctx, "message")

But that imposes a burden on all users, even those who don't need
the context.

Another solution is to create a new Logger each time a new trace
element ("span" in Open Telemetry terminology) is created, with an
attribute containing the span (or the context). This means that
everyone who creates a span must remember to create a Logger as well,
or it requires wrapping this package in another, context-aware one.

This CL implements an alternative: slog.FromContext(ctx) both
retrieves a Logger from ctx and arranges matters so that subsequent
uses of that Logger install ctx in the generated Record, where it is
available to Handlers.

The changes are:

- The Logger struct now includes a context as well as a Handler.

- FromContext(ctx) retrieves a Logger from ctx (or gets the default
  logger), as it always has. Then it creates and returns a new
  Logger that contains the same Handler as the retrieved one, but
  whose context is FromContext's argument, ctx.

- To avoid allocation in the above, Logger is changed from a pointer
  type to a value type.  Since it is immutable, its semantics are
  unchanged.

- Record has a new ctx field and a Context method to retrieve it.

- NewRecord gains an additional argument for the context.

- When a Logger l creates a Record r to pass to a Handler, it
  sets r.ctx to l.ctx.

The great advantage of this feature is that it delivers contexts to
Handlers "for free," with no calls other than the ones that would
normally take place: retrieving a Logger from a context and invoking a
logging method like Info or Log.

Some disadvantages:

- FromContext is overloaded to perform two conceptually unrelated
  tasks: fetch a Logger from a context, and deliver that context to a
  Handler. There is no a priori reason why a program that passes
  Loggers via contexts would need those contexts for anything else, or
  that a program that wants its Handler to see contexts must also pass
  Loggers around in contexts. But though these two properties need not
  be linked, they often are.  One way to look at it is that we treat a
  call to FromContext as a signal that Handlers want contexts.

- The context delivered to the Handler comes from the Logger, not the
  log line. If a new context is established between the call to
  FromContext and the call to the logging method, then the Handler
  will get the old context. We believe that this is an unlikely footgun:
  usually calls to FromContext will be close to logging calls, and
  we expect that often the two calls will be chained.

- Changing Logger to a value type quadruples its size (from a one-word
  pointer to a four-word struct). Our benchmarks show small slowdowns
  for the fastest operations (that make unrealistically optimistic
  assumptions), and none for slower (and more practical) ones.

Change-Id: I552b722c118d11406a05412e999c0b629b50228c