jsonrpc2: initialize Connection.async before calling binder.Bind
If Connection.async is not initialized before calling binder.Bind()
and a goroutine is spawned in binder.Bind() to monitor the end of the
connection using conn.Wait(), this c.Wait() can wait on a nil
c.async.ready which means it will never return.

    func (s *MyServer) Bind(
        ctx context.Context,
        conn *jsonrpc2.Connection,
    ) (jsonrpc2.ConnectionOptions, error) {
        // XXX register conn in *MyServer

        // monitor the end of the connection
        go func() {
            conn.Wait()
            conn.Close()
            // YYY do some cleanup in *MyServer
        }()
        return jsonrpc2.ConnectionOptions{}, nil
    }

The "// YYY do some cleanup in *MyServer" should be reached for each end
of connection.

If Connection.async is initialized too late, after conn.Wait() is
called in the Bind() goroutine, conn.Wait() never returns as it reads
on a nil channel.

Change-Id: I3c6e69d7dcc860479c4518ebd18e2431606322bc
GitHub-Last-Rev: 9067b3126969cdfaedbb0f5d03de5552975d1272
GitHub-Pull-Request: golang/exp#45