slog: use middle dot for TextHandler groups/scopes
Use a Unicode middle dot '·' as the separator between group/scope names
and keys in the TextHandler, instead of an ordinary dot '.'.

Middle dot is unlikely enough to occur in ordinary keys that we do not
feel obligated to escape it so that users can recover the group
structure from the key. If a user wants to do that, they can use
ReplaceAttr to do their own escaping.

Without escaping the group separator, it is easy to parse a key in the
TextHandler output:

    If you see a '"':
        Read to the next unescaped '"'
	Call strconv.Unquote
    else
        Read up to the '='

If we escaped group separators, parsing would be more complicated.

Change-Id: I6951ced8bb64dcb340c16372e0c27f49c861d88c