apidiff: account for receivers that are not named types
Until Go 1.20, the receiver type of a method that ultimately came from
a named type was also a named type. In 1.20, it is possible with
interface types that the receiver type is not named.

apidiff was using the type name to format a string in its output
messages.  Indeed, the entire message logic (messageset.go) assumed
that a types.Object had all the information needed to produce that
string. With Go 1.20, that is no longer the case.

This CL:

- Generalizes the messageset machinery to accept either a string or a
  types.Object.

- Determines a string for methods that uses the receiver type if it is
  named, as before, but falls back to the named type that the method
  originally came from.

Fixes golang/go#56162.

Change-Id: I1acfc3883d3668246e6e804eeb9fcb0e167959ab