slog: log errors correctly and efficiently
Logger.Error now avoids an append when adding the error Attr
to the Record. This also fixes the bug noticed in
https://github.com/golang/go/issues/56345#issuecomment-1303467620,
where a non-nil error and an omitted value combine to produce
incorrect output (see TestLoggerError).

Along the way, I noticed that the defaultHandler's reliance
on a fixed call depth is buggy, failing to output the right
source line when the handler is wrapped. I added a test, but
skipped it because I think the right solution must wait until
slog is in the standard library.

Change-Id: I301adfc646e4ec4688dbeb22cd27c0be8a3ca200