slices: add old build constraint syntax to help old go versions
When a project transitively depends on x/exp/slices,
go1.16 is confused and fails with errors like:

"//go:build comment without // +build comment"

Even when import of x/exp/slices guarded with build tags, so this
is imported only when compiling with go1.18, this error is unavoidable.
Let's wait a bit longer and help users who are stuck with old go.

Updates golang/go#51436

Change-Id: I7ba589644475f55585c11ba1f7650c5ede6bc637