slog: inform ReplaceAttr of groups
Formerly, HandleOptions.ReplaceAttr took only an Attr, so it was
ignorant of the groups that contained the Attr.

This change provides ReplaceAttr with a slice of currently open
groups along with the Attr they contain.

We use a sync.Pool to avoid allocating the group slice in the steady
state.

Also, resolve the Attr's value before passing it to ReplaceAttr.
We need to resolve to determine whether the Attr is a Group,
so let ReplaceAttr benefit from that work.

Change-Id: I9a50e5596f44f93534ce272f0e1681af529fdd94