slices: implement counting sort for 8 and 16 bits integer types
This is implemented using stack allocated counters and gives
O(N) speed.

Relevent benchmarks (quite noisy):
  goos: linux
  goarch: amd64
  pkg: golang.org/x/exp/slices
  cpu: AMD Ryzen 5 3600 6-Core Processor
  BenchmarkSlicesSortUint8_Random-12                   	   18021	     68122 ns/op
  BenchmarkSlicesSortUint8_Random_NotCounting-12       	     380	   2773817 ns/op
  BenchmarkSlicesSortUint16_Random-12                  	    2304	    527062 ns/op
  BenchmarkSlicesSortUint16_Random_NotCounting-12      	     201	   6120842 ns/op
  BenchmarkSlicesSortInt8_Random-12                    	   18991	     67556 ns/op
  BenchmarkSlicesSortInt8_Random_NotCounting-12        	     363	   2854590 ns/op
  BenchmarkSlicesSortInt16_Random-12                   	    2178	    535409 ns/op
  BenchmarkSlicesSortInt16_Random_NotCounting-12       	     198	   5883478 ns/op

We can see that for one million random elements:
- with 8 bits counting sort is ~41x faster. (2KiB counter array)
- with 16 bits counting sort is is ~12x faster. (0.5MiB counter array)

For sorted and reversed data counting sort is slower than with random data.
My understanding thx to the linux perf tool is that this is because with
random data the CPU is able to execute multiple instructions of the loop
in parallel and increment many counters at once, while with sorted and
reversed data there is contention because following iterations of the
loop increment the same memory address.
I guess a future optimisation for this would be to have some unrolled
loop that iterate over different part of the slice at once.

  BenchmarkSlicesSortUint8_Sorted-12                   	    6388	    188462 ns/op
  BenchmarkSlicesSortUint8_Sorted_NotCounting-12       	   22485	     49973 ns/op
  BenchmarkSlicesSortUint16_Sorted-12                  	    7528	    161340 ns/op
  BenchmarkSlicesSortUint16_Sorted_NotCounting-12      	   22443	     51712 ns/op
  BenchmarkSlicesSortInt8_Sorted-12                    	    6073	    204185 ns/op
  BenchmarkSlicesSortInt8_Sorted_NotCounting-12        	   23647	     51392 ns/op
  BenchmarkSlicesSortInt16_Sorted-12                   	    8822	    134943 ns/op
  BenchmarkSlicesSortInt16_Sorted_NotCounting-12       	   22784	     51640 ns/op

For sorted data pdqsort is ~4x faster, it might be intresting to try to
only use counting sort if pdqsort does not detect a sort array the first
time however the issue is that any time spent in pdqsort is time "wasted"
for counting sort since counting sort is not able to reuse pdqsort's progress.
This mean the sorted performance would be better, but the random performance
would get worst.

  BenchmarkSlicesSortUint8_Reversed-12                 	    6175	    196944 ns/op
  BenchmarkSlicesSortUint8_Reversed_NotCounting-12     	    2557	    472329 ns/op
  BenchmarkSlicesSortUint16_Reversed-12                	    7456	    160567 ns/op
  BenchmarkSlicesSortUint16_Reversed_NotCounting-12    	    3268	    366635 ns/op
  BenchmarkSlicesSortInt8_Reversed-12                  	    6082	    190537 ns/op
  BenchmarkSlicesSortInt8_Reversed_NotCounting-12      	    3500	    344764 ns/op
  BenchmarkSlicesSortInt16_Reversed-12                 	    8728	    135675 ns/op
  BenchmarkSlicesSortInt16_Reversed_NotCounting-12     	    3390	    353970 ns/op

Nothing intresting on reversed input, counting sort have similar perf as sorted
and pdqsort is faster than pdqsort on input but worst than pdqsort sort on
sorted.

Benchmarking around threshold suggest that, at least on my machine,
they are conservative for random inputs:
  BenchmarkCountingThreshold_Counting_Uint8-12         	 2526783	       514.1 ns/op
  BenchmarkCountingThreshold_NotCounting_Uint8-12      	 1074660	      1063 ns/op
  BenchmarkCountingThreshold_Counting_Int8-12          	 2519037	       505.4 ns/op
  BenchmarkCountingThreshold_NotCounting_Int8-12       	 1121792	      1004 ns/op

  BenchmarkCountingThreshold_Counting_Uint16-12        	    9660	    110817 ns/op
  BenchmarkCountingThreshold_NotCounting_Uint16-12     	    2629	    446313 ns/op
  BenchmarkCountingThreshold_Counting_Int16-12         	   11420	     98828 ns/op
  BenchmarkCountingThreshold_NotCounting_Int16-12      	    2785	    439846 ns/op

Change-Id: Ibaf79f67260859a9850fd7e9631a4f231f3a0061