exp/slog: add example for custom levels and names
The recommended way to introduce new `Level` (or rename an existing
one) is via `HandlerOptions.ReplaceAttr`. However, this is not
immediately intuitive to package users.

This introduces an example that adds additional log levels and gives
them custom names in a `ReplaceAttr` function. Given the number of users
who will want to introduce custom log levels, this feels like a good
addition to the library documentation.

Change-Id: I175440368fa1d4ce548d9c769ca22fd613151941