exp/slog: make tests more deterministic
This fixes an issue where test ordering could result in flakes because
of the dependency on the global logger. Multiple tests alter the state
of the default logger.

A better fix might be to not use the default logger (and just create our
own instance for the test), but I wanted to limit changes in the CL.

Change-Id: I01f16b5cb4ecd736f6345fca09e8f30ba1a58808