slog: correctly represent the zero Time in a Value
Storing the zero time.Time in a Value and then retrieving it
gave a non-zero time. The reason is that UnixNanos is called
to store a time.Time in a Value, and its result is undefined
for the zero time (or any time before the year 1678 or after 2262,
but I don't think we need to worry about those).

One solution, implemented in this CL, is to use a nil *time.Location
to represent the zero time. the time.Time.Location method never
returns nil, so a Value with a nil *time.Location in its `any` field
cannot be mistaken for any other Value, time.Time or otherwise.

Change-Id: Icba66f2ec16e21ed574c3336e1996ed759ab1c17