slog: Handlers drop empty groups.
If a group has no elements, then drop it completely instead of
outputting something like

	"g":{}

Besides removing clutter from the output, this allows LogValuers to
return a Value that is dropped.

Zap uses a special field type "Skip" for this purpose (equivalent to a
new slog value kind), but this solution is simpler.

Change-Id: I31885062660cd44e75349f6823d0e427b70e2f8e