slog: inline a group with an empty key
If a group Attr has an empty key, inline the group's members;
that is, do not qualify them by the group name.

This enables a LogValuer to insert more than one Attr into
the output. One case where that is useful is when the downstream
log processor does not support groups.

Change-Id: Iaf6b067ea4a0497717176d66857a318a513943f4