slog: change nopc variant from build tag to internal global
During benchmarking, we want to skip the call to runtime.Callers so we
can measure the speedup without it.

Previously we used the build tag `nopc` for this. But build tags are,
in a sense, exported API: anyone can rebuild with a particular tag. In
fact, at least one user believed that `nopc` was a supported feature
of slog.

This CL switches to using a global boolean variable. (The extra time
involved in checking the global is negligible.) To hide the variable
while letting both the slog and benchmarks packages access it,
we put it in an internal package.

Change-Id: I48c133e53c63cd36cc802bbd5f858b74cc19b9e8