slog: backport changes from stdlib
The standard library version of slog has diverged from this
version somewhat; bring them back into sync, as much as possible.

Since moving to the standard library, a Handler rule has changed: a
Handler should ignore the zero slog.Attr, but not an Attr whose key is
empty but whose value is non-zero. Change the check in slogtest to
reflect that.

Change-Id: Idd2084b84a3faf9b1796051af1578aee6558ed02