rand: add ChaCha8Source
DO NOT SUBMIT, this is just a benchmark demonstration. The current
assembly XORs the buffer with the keystream instead of overwriting,
which would be the correct and probably more performant thing to do.

The results are extremely sensitive to the design of the
micro-benchmark. Simply replacing the Source implementation in the old
BenchmarkSource yields on an M2

    goos: darwin
    goarch: arm64
    pkg: golang.org/x/exp/rand
            │   old.txt   │               new.txt               │
            │   sec/op    │   sec/op     vs base                │
    Source-8   3.098n ± 0%   4.458n ± 1%  +43.92% (p=0.000 n=10)

while the current benchmark (which does something with the result and
allows inlining, like performance sensitive applications would) looks
like this

    goos: darwin
    goarch: arm64
    pkg: golang.org/x/exp/rand
                    │  sink.txt   │
                    │   sec/op    │
    Source/PCG-8       2.148n ± 0%
    Source/ChaCha8-8   3.226n ± 0%

Interestingly, removing the sink but leaving the inlining makes PCG
*slower* and ChaCha8 faster 🤷‍♂️ I'll look at the generated code at
some point, but the conclusion is "slower, but in the same order of
magnitude".

    goos: darwin
    goarch: arm64
    pkg: golang.org/x/exp/rand
                    │ inline.txt  │
                    │   sec/op    │
    Source/PCG-8       2.463n ± 1%
    Source/ChaCha8-8   2.796n ± 1%

Change-Id: I8818753096f4aef14cace84a8974c3ecb5e9d0e2