x/exp/apidiff: add package path to change messages
Following the recent change to add module support to apidiff, it is now
no longer the case that all objects reported on in one apidiff execution
are in the same package. Therefore, add the package path, made relative
to the root of the apidiff execution, to object names when reporting
changes.

Fixes golang/go#61387

Change-Id: Idd748d46b68c0122f9f28ea3b51cc0519448b672