jsonrpc2: do not panic when sending unmarshalable params
If params given to (*Connection).Call is not JSON-marshalable, Call
returns immediately an *AsyncCall instance containing the
corresponding error, but its ctx field is not initialized. This
produces a panic later in golang.org/x/exp/event.New, as it does not
expect a nil context.Context.

Fixes golang/go#61654.

Change-Id: I6464ab5fde85670267b25821e0b57af04c331c8c