Migrated `git status --porcelain` to `git diff`
The expectation of this application is that it is to be utilized in release pipelines. Due to this expectation, pipelines utilize volume mounting to transfer artifacts (untracked files) across different jobs (specfically referring to gitlab here). This is a problem because git status --porcelain checks for untracked files and returns them as a list. This causes this pipeline to fail unexpectedly. In testing, `git reset --hard` will not remove these, while a `git clean -fdx` will.

I would recommend that in a follow-up commit there is an update to the available args for this cmd to control if untracked files are to be included or not (eg: --untracked)

Change-Id: I819ad41795bcd355b2bf42d57c050d466ea52876
GitHub-Last-Rev: e55eb3d9837170a2b08b5367ef7fb3bf545624ba
GitHub-Pull-Request: golang/exp#72