trace: update gen.bash for oldtrace and make it work seamlessly
Currently gen.bash doesn't work out-of-the-box because it doesn't
exactly replace contents in the trace directory in some cases. It also
may end up mutating the flight recorder files, which isn't great.

This change adds a special case for the flight recorder files, adds
support for oldtrace (which will be imported shortly), and also switches
to using rsync, which is more robust than cp for the initial copy (and
should hopefully capture file deletions correctly as well).

Change-Id: I3a808387a25c787a1edb70fcff1c95fb5649baeb