trace: support BSD version of sed in gen.bash
The GNU version of sed (common on Linux) will modify files in place when
it sees the -i flag with no argument at all, as in [sed -i -e '...'].
The BSD version of sed (common on macOS) will modify files in place when
it's given an explicit zero-length argument, as in [sed -i '' -e '...'].

But they agree on at least one way of specifying a non-zero-length
backup suffix, [sed -i '.tmp' -e '...']. Use that and then delete the
backups.

In addition, change the expressions we pass to sed to make them a bit
more explicit and to require less escaping.

Fixes golang/go#68403

Change-Id: Icfc3863f4abe26bc77d9c0bc3b778067db6387de